requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudinary==1.25.0
dj-database-url==0.5.0
Django==3.1.7
django-cloudinary-storage==0.3.0
django-heroku==0.0.0
djangorestframework==3.12.2
gunicorn==20.1.0
idna==2.10
mypy-extensions==0.4.3
numpy==1.20.1
pathspec==0.8.1
Pillow==8.1.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
